backend/requirements.txt
Flask==2.3.5
flask-cors==4.2.0
requests==2.31.0
python-dotenv==1.0.1